+++ openstl/methods/predrnn.py
@@ -1,11 +1,11 @@
+import time
 import torch
 import torch.nn as nn
-import numpy as np
 from timm.utils import AverageMeter
 from tqdm import tqdm
 
 from openstl.models import PredRNN_Model
-from openstl.utils import (reshape_patch, reshape_patch_back,
+from openstl.utils import (reduce_tensor, reshape_patch, reshape_patch_back,
                            reserve_schedule_sampling_exp, schedule_sampling)
 from .base_method import Base_method
 
@@ -29,8 +29,8 @@ class PredRNN(Base_method):
         num_layers = len(num_hidden)
         return PredRNN_Model(num_layers, num_hidden, args).to(self.device)
 
-    def _predict(self, batch_x, batch_y):
-        """Forward the model."""
+    def _predict(self, batch_x, batch_y, **kwargs):
+        """Forward the model"""
         # reverse schedule sampling
         if self.args.reverse_scheduled_sampling == 1:
             mask_input = 1
@@ -53,32 +53,28 @@ class PredRNN(Base_method):
         if self.args.reverse_scheduled_sampling == 1:
             real_input_flag[:, :self.args.pre_seq_length - 1, :, :] = 1.0
 
-        img_gen, _ = self.model(test_dat, real_input_flag)
+        img_gen, _ = self.model(test_dat, real_input_flag, return_loss=False)
         img_gen = reshape_patch_back(img_gen, self.args.patch_size)
         pred_y = img_gen[:, -self.args.aft_seq_length:].permute(0, 1, 4, 2, 3).contiguous()
 
         return pred_y
 
-    def forward_test(self, batch_x, batch_y):
-        """Evaluate the model"""
-        batch_x, batch_y = batch_x.to(self.device), batch_y.to(self.device)
-
-        with self.amp_autocast():
-            pred_y = self._predict(batch_x, batch_y)
-
-        return dict(zip(['inputs', 'preds', 'trues'],
-                        [batch_x, pred_y, batch_y]))
-
-    def train_one_epoch(self, runner, train_loader, epoch, num_updates, loss_mean, eta, **kwargs):
+    def train_one_epoch(self, runner, train_loader, epoch, num_updates, eta=None, **kwargs):
+        """Train the model with train_loader."""
+        data_time_m = AverageMeter()
         losses_m = AverageMeter()
         self.model.train()
         if self.by_epoch:
             self.scheduler.step(epoch)
+        train_pbar = tqdm(train_loader) if self.rank == 0 else train_loader
 
-        train_pbar = tqdm(train_loader)
+        end = time.time()
         for batch_x, batch_y in train_pbar:
+            data_time_m.update(time.time() - end)
             self.model_optim.zero_grad()
+
             batch_x, batch_y = batch_x.to(self.device), batch_y.to(self.device)
+            runner.call_hook('before_train_iter')
 
             # preprocess
             ims = torch.cat([batch_x, batch_y], dim=1).permute(0, 1, 3, 4, 2).contiguous()
@@ -90,19 +86,43 @@ class PredRNN(Base_method):
                 eta, real_input_flag = schedule_sampling(
                     eta, num_updates, ims.shape[0], self.args)
 
-            img_gen, loss = self.model(ims, real_input_flag)
-            loss.backward()
+            with self.amp_autocast():
+                img_gen, loss = self.model(ims, real_input_flag)
+
+            if not self.dist:
+                losses_m.update(loss.item(), batch_x.size(0))
+
+            if self.loss_scaler is not None:
+                if torch.any(torch.isnan(loss)) or torch.any(torch.isinf(loss)):
+                    raise ValueError("Inf or nan loss value. Please use fp32 training!")
+                self.loss_scaler(
+                    loss, self.model_optim,
+                    clip_grad=self.args.clip_grad, clip_mode=self.args.clip_mode,
+                    parameters=self.model.parameters())
+            else:
+                loss.backward()
+                self.clip_grads(self.model.parameters())
+
             self.model_optim.step()
-            if not self.by_epoch:
-                self.scheduler.step(epoch)
-
+            torch.cuda.synchronize()
             num_updates += 1
-            loss_mean += loss.item()
-            losses_m.update(loss.item(), batch_x.size(0))
 
-            train_pbar.set_description('train loss: {:.4f}'.format(loss.item()))
+            if self.dist:
+                losses_m.update(reduce_tensor(loss), batch_x.size(0))
+
+            if not self.by_epoch:
+                self.scheduler.step()
+            runner.call_hook('after_train_iter')
+            runner._iter += 1
+
+            if self.rank == 0:
+                log_buffer = 'train loss: {:.4f}'.format(loss.item())
+                log_buffer += ' | data time: {:.4f}'.format(data_time_m.avg)
+                train_pbar.set_description(log_buffer)
+
+            end = time.time()  # end for
 
         if hasattr(self.model_optim, 'sync_lookahead'):
             self.model_optim.sync_lookahead()
 
-        return num_updates, loss_mean, eta
+        return num_updates, losses_m, eta