+++ openstl/methods/simvp.py
@@ -27,7 +27,7 @@ class SimVP(Base_method):
         return SimVP_Model(**config).to(self.device)
 
     def _predict(self, batch_x, batch_y=None, **kwargs):
-        """Forward the model."""
+        """Forward the model"""
         if self.args.aft_seq_length == self.args.pre_seq_length:
             pred_y = self.model(batch_x)
         elif self.args.aft_seq_length < self.args.pre_seq_length:
@@ -76,7 +76,7 @@ class SimVP(Base_method):
 
             if self.loss_scaler is not None:
                 if torch.any(torch.isnan(loss)) or torch.any(torch.isinf(loss)):
-                    raise ValueError("Inf or nan loss value. Please use fp32 training instead!")
+                    raise ValueError("Inf or nan loss value. Please use fp32 training!")
                 self.loss_scaler(
                     loss, self.model_optim,
                     clip_grad=self.args.clip_grad, clip_mode=self.args.clip_mode,