+++ openstl/models/mau_model.py
@@ -105,7 +105,7 @@ class MAU_Model(nn.Module):
         self.conv_last_sr = nn.Conv2d(
             self.frame_channel * 2, self.frame_channel, kernel_size=1, stride=1, padding=0)
 
-    def forward(self, frames_tensor, mask_true):
+    def forward(self, frames_tensor, mask_true, **kwargs):
         # [batch, length, height, width, channel] -> [batch, length, channel, height, width]
         frames = frames_tensor.permute(0, 1, 4, 2, 3).contiguous()
         mask_true = mask_true.permute(0, 1, 4, 2, 3).contiguous()
@@ -161,7 +161,6 @@ class MAU_Model(nn.Module):
                 T_pre[i].append(T_t[i])
             out = S_t
 
-
             for i in range(len(self.decoders)):
                 out = self.decoders[i](out)
                 if self.configs.model_mode == 'recall':
@@ -172,6 +171,9 @@ class MAU_Model(nn.Module):
         
         # [length, batch, channel, height, width] -> [batch, length, height, width, channel]
         next_frames = torch.stack(next_frames, dim=0).permute(1, 0, 2, 3, 4).contiguous()
-        loss = self.MSE_criterion(next_frames, frames[:, 1:])
+        if kwargs.get('return_loss', True):
+            loss = self.MSE_criterion(next_frames, frames[:, 1:])
+        else:
+            loss = None
 
         return next_frames, loss