+++ openstl/utils/parser.py
@@ -26,7 +26,9 @@ def create_parser():
     parser.add_argument('--fps', action='store_true', default=False,
                         help='Whether to measure inference speed (FPS)')
     parser.add_argument('--empty_cache', action='store_true', default=True,
-                        help='Whether to empty cuda cache after training')
+                        help='Whether to empty cuda cache after GPU training')
+    parser.add_argument('--find_unused_parameters', action='store_true', default=False,
+                        help='Whether to find unused parameters in forward during DDP training')
     parser.add_argument('--resume_from', type=str, default=None, help='the checkpoint file to resume from')
     parser.add_argument('--auto_resume', action='store_true', default=False,
                         help='When training was interupted, resume from the latest checkpoint')